bot/requirements.txt
python-telegram-bot==20.0a4
requests==2.28.1